refactor(app): split orchestrator into user/ and session/ subdirs
